+++ src/components/frame/v5/pages/BalancePage/partials/TokenAvatar/types.ts
@@ -2,6 +2,6 @@ import { NativeTokenStatus, TokenFragment } from '~gql';
 
 export interface TokenAvatarProps {
   token: TokenFragment;
-  tokenAddress: string;
-  nativeTokenStatus: NativeTokenStatus;
+  isTokenNative: boolean;
+  nativeTokenStatus?: NativeTokenStatus;
 }
