+++ src/components/v5/common/ActionSidebar/partials/forms/ManagePermissionsForm/partials/PermissionsTable/hooks.tsx
@@ -5,7 +5,7 @@ import React, { useMemo } from 'react';
 import { UserRole, USER_ROLE } from '~constants/permissions';
 import { formatText } from '~utils/intl';
 import { FormSwitch } from '~v5/common/Fields/Switch';
-import { TableProps } from '~v5/common/Table/types';
+import { TableProps } from '~v5/common/TableOld/types';
 
 import { PERMISSIONS_TABLE_CONTENT } from './consts';
 import { CustomPermissionTableModel, PermissionsTableModel } from './types';
