+++ src/components/v5/common/TableOld/index.ts
@@ -1,0 +1,1 @@
+export { default } from './TableOld';
