+++ src/components/v5/common/TableWithMeatballMenu/types.ts
@@ -2,7 +2,7 @@ import { Row } from '@tanstack/react-table';
 
 import { MeatBallMenuProps } from '~v5/shared/MeatBallMenu/types';
 
-import { TableProps } from '../Table/types';
+import { TableProps } from '../TableOld/types';
 
 export interface TableWithMeatballMenuProps<T> extends TableProps<T> {
   getMenuProps: (row: Row<T>) => MeatBallMenuProps | undefined;
